add pipeline script

+++ openie_freq.py
@@ -1,4 +1,5 @@
 
+import os
 import sys
 import json
 import numpy as np
@@ -11,9 +12,13 @@ from misc.openie_utils import remap_preds, test_caps, post_process_preds, post_p
 
 from config import VG_SGG_DICT_FN
 
-TRIPLES_FILE = "descriptions_svo.txt"
-INTM_FILE = "descriptions_text.npy"
-OUTPUT_FILE = "descriptions_freq.npy"
+TRIPLES_FILE = sys.argv[1]
+INTM_FILE = sys.argv[2]
+OUTPUT_FILE = sys.argv[3]
+
+# TRIPLES_FILE = "descriptions_svo.txt"
+# INTM_FILE = "descriptions_text.npy"
+# OUTPUT_FILE = "descriptions_freq.npy"
 
 # split = sys.argv[1]
 # assert split in ["a", "b", "c", "d", "e", "f"]
@@ -45,35 +50,43 @@ preds, predicate_to_ind = remap_preds(predicate_to_ind)
 num_classes = len(objs)
 num_predicates = len(preds)
 
-triples = []
-print("Reading file...")
-with codecs.open(TRIPLES_FILE) as f:
-    for line in f:
-        svo = [x.strip() for x in line.split("|")]
-        if len(svo) != 3:
-            continue
-        triples.extend(svo)
-print("Done.")
+if not os.path.isfile(INTM_FILE):
 
-# lemmatizing
-import spacy
-spacy.prefer_gpu()
+    print("Generating file with lemmatized text:", INTM_FILE)
 
-nlp = spacy.load('en', disable=['parser', 'ner'])
+    triples = []
+    print("Reading file...")
+    with codecs.open(TRIPLES_FILE) as f:
+        for line in f:
+            svo = [x.strip() for x in line.split("|")]
+            if len(svo) != 3:
+                continue
+            triples.extend(svo)
+    print("Done.")
 
-_triples = triples
+    # lemmatizing
+    import spacy
+    spacy.prefer_gpu()
 
-ltriples = []
-for doc in tqdm(nlp.pipe(
-    texts=_triples, 
-    as_tuples=False, 
-    n_threads=2, 
-    batch_size=10000), total=len(_triples)):    
-    ltriples.append([x.lemma_ for x in doc])
-ltriples = np.array(ltriples)
+    nlp = spacy.load('en', disable=['parser', 'ner'])
 
-ltriples = np.reshape(ltriples, (-1, 3))
-np.save(INTM_FILE, ltriples)
+    _triples = triples
+
+    ltriples = []
+    for doc in tqdm(nlp.pipe(
+        texts=_triples, 
+        as_tuples=False, 
+        n_threads=2, 
+        batch_size=10000), total=len(_triples)):    
+        ltriples.append([x.lemma_ for x in doc])
+    ltriples = np.array(ltriples)
+
+    ltriples = np.reshape(ltriples, (-1, 3))
+    np.save(INTM_FILE, ltriples)
+
+else:
+
+    print("Loading file with lemmatized text from disk", INTM_FILE)
 
 # actual build of prior
 ltriples = np.load(INTM_FILE)
